[REF] l10n_it_asset_management: Extract common tests setup
